uint8 is ugly; changed back to uint16

+++ combineAndDivideTifs.m
@@ -17,7 +17,7 @@ try
     disp('Reading in tif files:');
     for i=1:length(file)
         disp(i);
-        movie=tiffRead([pathname file{i}],'uint8');
+        movie=tiffRead([pathname file{i}],'uint16');
         combinedMovie=cat(3,combinedMovie,movie);
     end
 catch
@@ -145,14 +145,14 @@ if ~isempty(combinedMovie)
     end
     try
         for i=1:length(combined)
-            tiffWrite(combined{i},['combined' num2str(i) '.tif'],pathname,'uint8');
+            tiffWrite(combined{i},['combined' num2str(i) '.tif'],pathname,'uint16');
         end
     catch
         % Sometimes, disk access fails due to intermittent
         % network problem. In that case, wait and re-try once:
         pause(60);
         for i=1:length(combined)
-            tiffWrite(combined{i},['combined' num2str(i) '.tif'],pathname,'uint8');
+            tiffWrite(combined{i},['combined' num2str(i) '.tif'],pathname,'uint16');
         end
     end
 else
